backend/requirements.txt
# backend/requirements.txt
Flask>=2.0
google-generativeai
Pillow
gunicorn
Flask-Cors